Fix bug with ispo() function, incorrect use of the chores flag
Function ispo() replaced with code from the ispo2() function.

Fixes an issue for goods only instances:
Fixed a bug which caused the ILP within the function to be infeasible for goods only instances.
The bug is due to incorrect use of the chores (True/False) input variable which is set to False for goods instances.
This causes the upper bound for the utility of agents to be set to 0 while the lower bound is typically some positive integer.
The upper bound is set to 0 because the desired upper bound is multiplied by the value of the chores variable setting it to 0. The upper bound should be set to the budget (or more precisely, the total value of all goods for the agent).

The issue seems to have been introduced due to poor versioning earlier causing unfinished code to be committed.